requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
pillow
transformers>=4.42.0
accelerate
bitsandbytes
timm
einops
safetensors
sentencepiece
protobuf
requests
